Reduced SQL queries to one, simplified data collection
Refactored the SQL code to return the results in one query and to hold in a dataframe for the remainder of the program. Simplified the code instead of relying on data frames for each and every step, instead storing vendor data in nested dictionaries that then could be unpacked as needed. More efficient.
This also simplified the code defining each French Vendor's vendor contacts and CS contacts.

Signed-off-by: Ryan Buzar <109171271+RyanBuzar@users.noreply.github.com>